chore: add file
add readme and screenshots folder

+++ src/components/Article/manage.tsx
@@ -1,6 +1,13 @@
 import * as React from 'react'
 import * as styles from './style.css'
-import { Table, Button, Image, Form } from 'react-bootstrap'
+import {
+	Table,
+	Button,
+	Image,
+	Form,
+	DropdownButton,
+	Dropdown,
+} from 'react-bootstrap'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import {
 	faTags,
@@ -10,10 +17,22 @@ import {
 	faDesktop,
 	faSatelliteDish,
 } from '@fortawesome/free-solid-svg-icons'
-import { ArticleModel, ArticleDataModel, TagDataModel, CategoryDataModel } from '@app/store/models'
+import {
+	ArticleModel,
+	ArticleDataModel,
+	TagDataModel,
+	CategoryDataModel,
+} from '@app/store/models'
 import { ArticleActions } from '@app/store/actions'
 import { formatDate } from '@app/utils'
-import { ConfirmModal, Paging, Search, Selector } from '@app/components'
+import {
+	Notication,
+	ConfirmModal,
+	Paging,
+	Search,
+	Selector,
+} from '@app/components'
+import { INotice } from '@app/interfaces/notice'
 import * as CONFIG from '@app/config'
 
 const STATE_LIST = [
@@ -48,8 +67,8 @@ const artHeads = [
 	'围观',
 	'评论',
 	'访问',
-	'状态',
 	'类型',
+	'状态',
 	'时间',
 	'操作',
 ]
@@ -71,6 +90,11 @@ export namespace Article {
 		keyword: string
 		querys: any
 		currentPage: number
+		show: boolean
+		type: string
+		content: string
+		checkedValues: string[]
+		deleteType: string
 	}
 }
 
@@ -84,35 +108,78 @@ export class Article extends React.Component<Article.IProps, Article.IState> {
 			keyword: '',
 			querys: {},
 			currentPage: 1,
+			show: false,
+			type: 'info',
+			content: '添加成功',
+			checkedValues: [],
+			deleteType: 'single'
 		}
 
+		this.closeModal = this.closeModal.bind(this)
+		this.handleConfirm = this.handleConfirm.bind(this)
+		this.handleDelete = this.handleDelete.bind(this)
+		this.handleChecked = this.handleChecked.bind(this)
 		this.handleCategory = this.handleCategory.bind(this)
 		this.handlePagination = this.handlePagination.bind(this)
 		this.handleKeywordSearch = this.handleKeywordSearch.bind(this)
 	}
 
-	openModal(id: string) {
+	showNotice(obj: INotice) {
+		const { type, content } = obj
 		this.setState({
-			articleId: id,
+			show: true,
+			type,
+			content,
+		})
+	}
+
+	openModal() {
+		this.setState({
 			showModal: true,
 		})
 	}
 
-	handleDelete(e: React.MouseEvent<HTMLButtonElement>) {
-		this.props.deleteArticle(this.state.articleId)
+	closeModal() {
 		this.setState({
-			articleId: '',
 			showModal: false,
+			deleteType: 'single',
 		})
 	}
+
+	// 删除
+	handleConfirm() {
+		const { deleteType } = this.state
+
+		if (deleteType === 'single') {
+			// 单个删除
+			this.props.deleteArticle(this.state.articleId)
+		}
+		if (deleteType === 'batchPublish') {
+			// 批量发布
+			console.log('batch publish', this.state.checkedValues)
+		}
+		if (deleteType === 'batchDelete') {
+			// 批量删除
+			console.log('batch delete', this.state.checkedValues)
+		}
+		
+		this.setState({
+			showModal: false,
+			deleteType: 'single',
+		})
+		// this.showNotice({ type: 'success', content: '删除成功！' })
+	}
+
 	// 更新
 	handleUpdate(_id: string) {
 		this.props.editArticle(_id)
 	}
+
 	// 预览
 	handleView(_id: string) {
 		window.open(`${CONFIG.APP.baseUrl}article/${_id}`)
 	}
+
 	// 发布
 	handlePublish(article: ArticleModel) {
 		const { _id } = article
@@ -147,7 +214,7 @@ export class Article extends React.Component<Article.IProps, Article.IState> {
 		querys['keyword'] = keyword
 		querys['page'] = currentPage
 		querys['isAuthenticated'] = true
-		
+
 		if (value !== '') {
 			querys[type] = value
 		} else {
@@ -155,39 +222,92 @@ export class Article extends React.Component<Article.IProps, Article.IState> {
 		}
 
 		this.setState({
-			querys
+			querys,
 		})
 
 		this.props.getArticleList(querys)
 	}
 
+	// 复选框选择
+	handleChecked(_id: any, event: React.ChangeEvent<HTMLInputElement>) {
+		const { checked } = event.target
+		let { checkedValues } = this.state
+		
+		if (checked && checkedValues!.filter((item: any) => item !== _id)) {
+			checkedValues!.push(_id)
+		} else {
+			checkedValues = checkedValues!.filter((item: any) => item !== _id)
+		}
+		console.log('checked', checkedValues)
+		this.setState({ checkedValues })
+	}
+	
+	// 单个删除
+	handleDelete(id: string, event: React.ChangeEvent<HTMLButtonElement>) {
+		this.openModal()
+		
+		this.setState({
+			articleId: id,
+		})
+	}
+
+	// 批量操作
+	handleBatch(name: string, event: React.ChangeEvent<HTMLButtonElement>) {
+		this.openModal()
+
+		this.setState({
+			deleteType: name
+		})
+	}
+
 	renderHeaderBar(): JSX.Element | void {
 		const { categories, tags, articles } = this.props
 		const { currentPage } = this.state
-		const catSelectList = [{_id: '', name: '分类不限'}, ...categories.data]
-		const tagSelectList = [{_id: '', name: '标签不限'}, ...tags.data]
-		
+		const catSelectList = [{ _id: '', name: '分类不限' }, ...categories.data]
+		const tagSelectList = [{ _id: '', name: '标签不限' }, ...tags.data]
+
 		return (
 			<div className="flex pdb10 pdt10">
 				<div className="flex flex30">
 					<Search
+						
 						placeholder="搜索关键字"
 						handleChange={(name: string, val: any) =>
 							this.keywordChange(name, val)
 						}
 						handleSearch={this.handleKeywordSearch}
 					/>
+					<span className="mr10"></span>
+					<DropdownButton id="dropdown-basic-button" variant="warning" title="批量操作">
+						<Dropdown.Item onClick={(e: any) => this.handleBatch("batchPublish", e)}>发布</Dropdown.Item>
+						<Dropdown.Item onClick={(e: any) => this.handleBatch("batchDelete", e)}>删除</Dropdown.Item>
+					</DropdownButton>
 				</div>
 				<div className="flex flex1 flex-end">
 					<div className={styles.filterBox}>
-						<Selector list={catSelectList} onChange={(e: any) => this.handleCategory(e, 'category')}/>
-						<Selector list={tagSelectList} onChange={(e: any) => this.handleCategory(e, 'tag')}/>
-						<Selector list={STATE_LIST} onChange={(e: any) => this.handleCategory(e, 'state')}/>
-						<Selector list={ORIGIN_LIST} onChange={(e: any) => this.handleCategory(e, 'origin')}/>
-						<Selector list={PUBLISH_LIST} onChange={(e: any) => this.handleCategory(e, 'public')}/>
+						<Selector
+							list={catSelectList}
+							onChange={(e: any) => this.handleCategory(e, 'category')}
+						/>
+						<Selector
+							list={tagSelectList}
+							onChange={(e: any) => this.handleCategory(e, 'tag')}
+						/>
+						<Selector
+							list={STATE_LIST}
+							onChange={(e: any) => this.handleCategory(e, 'state')}
+						/>
+						<Selector
+							list={ORIGIN_LIST}
+							onChange={(e: any) => this.handleCategory(e, 'origin')}
+						/>
+						<Selector
+							list={PUBLISH_LIST}
+							onChange={(e: any) => this.handleCategory(e, 'public')}
+						/>
 					</div>
 					<Paging
-						total={articles.pagination.total}
+						total={articles.pagination && articles.pagination.total}
 						active={currentPage}
 						handlePage={this.handlePagination}
 					/>
@@ -224,7 +344,7 @@ export class Article extends React.Component<Article.IProps, Article.IState> {
 				{data.map((it: any) => (
 					<tr key={it._id}>
 						<td>
-    					<Form.Check type="checkbox" label={it.id} />
+							<Form.Check type="checkbox" label={it.id} onChange={(e: any) => this.handleChecked(it._id, e)} />
 						</td>
 						<td className={styles.thumbnail}>
 							<Image
@@ -239,7 +359,11 @@ export class Article extends React.Component<Article.IProps, Article.IState> {
 							{it.category.length > 0 &&
 								it.category.map((cate: any, idx: number) => (
 									<div key={cate._id}>
-										<FontAwesomeIcon icon={faFolder} size="1x" />
+										<FontAwesomeIcon
+											icon={faFolder}
+											size="1x"
+											color="#727374"
+										/>
 										<span> {cate.name} </span>
 									</div>
 								))}
@@ -248,14 +372,18 @@ export class Article extends React.Component<Article.IProps, Article.IState> {
 							{it.tag.length > 0 &&
 								it.tag.map((it: any, idx: number) => (
 									<div key={it._id}>
-										<FontAwesomeIcon icon={faTags} size="1x" />
+										<FontAwesomeIcon icon={faTags} size="1x" color="#727374" />
 										<span> {it.name} </span>
 									</div>
 								))}
 						</td>
 						<td>{it.meta.likes === 0 ? '无人鼓励' : it.meta.likes + '个赞'}</td>
 						<td>{it.meta.views === 0 ? '无人围观' : it.meta.views + '围观'}</td>
-						<td>{it.meta.comments === 0 ? '无人说话' : it.meta.comments + '条评论' }</td>
+						<td>
+							{it.meta.comments === 0
+								? '无人说话'
+								: it.meta.comments + '条评论'}
+						</td>
 						<td>
 							{it.state === ArticleModel.EStatePublic.Public
 								? '公开'
@@ -303,22 +431,24 @@ export class Article extends React.Component<Article.IProps, Article.IState> {
 								修改文章
 							</Button>
 
-							<Button
-								size="sm"
-								variant="success"
-								onClick={() => this.handlePublish(it)}>
-								<FontAwesomeIcon
-									icon={faSatelliteDish}
-									size="1x"
-									style={{ marginRight: '2px' }}
-								/>
-								发布文章
-							</Button>
+							{it.origin !== ArticleModel.EStateOrigin.Reprint ? (
+								<Button
+									size="sm"
+									variant="success"
+									onClick={() => this.handlePublish(it)}>
+									<FontAwesomeIcon
+										icon={faSatelliteDish}
+										size="1x"
+										style={{ marginRight: '2px' }}
+									/>
+									发布文章
+								</Button>
+							) : null}
 
 							<Button
 								size="sm"
 								variant="warning"
-								onClick={() => this.openModal(it._id)}>
+								onClick={(e: any) => this.handleDelete(it._id, e)}>
 								<FontAwesomeIcon
 									icon={faTrash}
 									size="1x"
@@ -351,13 +481,22 @@ export class Article extends React.Component<Article.IProps, Article.IState> {
 	}
 
 	render() {
-		const { showModal } = this.state
+		const { showModal, show, type, content } = this.state
 		return (
 			<div className="module">
+				<Notication
+					show={show}
+					type={type}
+					content={content}
+					onClose={() => {
+						this.setState({ show: false })
+					}}
+					autohide
+				/>
 				<ConfirmModal
 					show={showModal}
-					onHide={() => this.setState({ showModal: false })}
-					onClose={(e: any) => this.handleDelete(e)}
+					onHide={this.closeModal}
+					onClose={this.handleConfirm}
 				/>
 				{this.renderArticleList()}
 			</div>
